dagger gem
allow to create a real gem so that it can be used with a dagger run

Gem is called 'dagger-sdk' because 'dagger' and 'ruby-dagger' already
exist.

To test:

- generate ruby code:
  - dagger -m .dagger call sdk ruby generate -o .
- build the gem:
  - cd sdk/ruby; gem build dagger.gemspec
- install it locally:
  - gem install dagger-sdk-0.0.0.gem

This will allow to have a local version of the gem, without the need to
deploy it.
Once done, you can follow all the steps detailed in the sdk/ruby/README.md
